Update Project.xml
This will update Psych Engine's location to a new location like it's seen in Kade Engine instead of using the usual route like it's in the original game. This of course will get the progress of the players deleted since it'll be in a new path. Including the window's title to be called Friday Night Funkin': Psych Engine instead of Friday Night Funkin'.