Add Sarvente's Mid-Fight Masses, but currently horrible (read commit)
Charts are horrible, like they've always been in this mod.
Animations are currently a bit janky and I will fix this in the future.
Stages is just lazy ported, so no animations or anything.

Rechart of these songs will be done in the future. But atleast the files are here now :)