Made it changeable through events, script or whatever
With this, the auto zoom on beat value, can be changed using events, scripts or whatever (NOT in the json chart file).
This makes it easier to, for example, make the camera (in a song where the auto zoom value is true) doesn't have auto zooms on beat in a little part of the song.