Custom Default Difficulty and Fallback
This Pull Request does a few things in relation to a default difficulty.

The first is that you can change the week's default difficulty by adding to your week json. For example based off the Crazy difficulty example in the wiki, `"defaultDifficulty": "Crazy"`. This means when in Story Mode or Freeplay when selecting a week or song, instead of starting at "Normal", it starts at "Crazy". This can also be changed in the Week Editor.

The second thing this does is make it so you have a default chart that other difficulties can use if it doesn't have a unique chart. It works the same as if the base game difficulty is "Normal". This can be seen with Tutorial which I've removed the Easy difficulty. Since that chart no longer exists, it takes from `tutorial.json` instead of `tutorial-easy.json`. This is useful for modders who make songs that have multiple difficulties, but the same chart.